requirements.txt
adminui==1.5.1
aiofile==3.8.8
alembic==1.12.1
anyio==4.0.0
bidict==0.22.1
blinker==1.6.3
caio==0.9.13
certifi==2023.7.22
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
fastapi==0.96.1
Flask==3.0.0
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pipdeptree==2.13.0
platformdirs==3.11.0
psutil==5.9.6
pydantic==1.10.13
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-engineio==4.8.0
python-multipart==0.0.5
python-socketio==5.10.0
pytz==2023.3.post1
redis==4.6.0
rich==13.6.0
setuptools==68.2.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
starlette-admin==0.9.0
tabulate==0.9.0
typer==0.4.2
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
websockets==10.4
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.2.0
